+++ src/3D/scene/scene.go
@@ -8,5 +8,5 @@ import (
 
 type Scene interface {
 	GetMaterial(p vector.Vec3) primitives.Material
-	Trace(r vector.Ray3D, tmax float64) (visibility float64, c color.Color)
+	Trace(r vector.Ray3D, tmax float32) (visibility float32, c color.Color)
 }
