+++ src/libs/vector/vector2D.go
@@ -1,23 +1,23 @@
 package vector
 
-import "math"
+import math "github.com/chewxy/math32"
 
 type Vec2 struct {
-	X, Y float64
+	X, Y float32
 }
 
-func NewVec2fromAngle(angle float64) Vec2 {
+func NewVec2fromAngle(angle float32) Vec2 {
 	return Vec2{
 		X: math.Cos(angle),
 		Y: math.Sin(angle),
 	}
 }
 
-func (v *Vec2) Length() float64 {
+func (v *Vec2) Length() float32 {
 	return math.Hypot(v.X, v.Y)
 }
 
-func (v *Vec2) Normalize() float64 {
+func (v *Vec2) Normalize() float32 {
 	l := v.Length()
 	v.X /= l
 	v.Y /= l
@@ -40,7 +40,7 @@ type Ray2D struct {
 	Dir Vec2
 }
 
-func (r *Ray2D) Trace(t float64) Vec2 {
+func (r *Ray2D) Trace(t float32) Vec2 {
 	// Move the ray's point along its direction vector by distance d
 	return Vec2{
 		r.P.X + r.Dir.X*t,
